# apps/web/lib/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunRequest": {
        "description": "Ein neuer Lauf. Führt bewusst keine ``user_id``.",
        "properties": {
          "channel": {
            "default": "text",
            "enum": [
              "text",
              "voice"
            ],
            "title": "Channel",
            "type": "string"
          },
          "input": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Input",
            "type": "string"
          }
        },
        "required": [
          "input"
        ],
        "title": "RunRequest",
        "type": "object"
      },
      "RunView": {
        "description": "Sicht auf einen Lauf.\n\nBewusst nicht das Vertragsmodell selbst: ``Run`` führt ``user_id``, und ein\nAntwortmodell, das eine Identität trägt, ist ein Antwortmodell, das morgen\njemand als Request-Modell wiederverwendet. Der Nutzer steht ohnehin fest —\ner hat gefragt.",
        "properties": {
          "data_class": {
            "title": "Data Class",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "goal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "intent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent"
          },
          "is_multi_step": {
            "title": "Is Multi Step",
            "type": "boolean"
          },
          "plan": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PlanStepView"
            },
            "title": "Plan",
            "type": "array"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "taint_level": {
            "title": "Taint Level",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          },
          "trigger": {
            "title": "Trigger",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "trigger",
          "taint_level",
          "data_class",
          "intent",
          "is_multi_step",
          "trace_id",
          "started_at",
          "finished_at"
        ],
        "title": "RunView",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Alle Endpunkte außer der Erstinbetriebnahme und dem Anmelde-Handshake verlangen eine gültige Sitzung. Die Identität stammt ausschließlich aus dieser Sitzung, nie aus Angaben im Request.",
    "summary": "Persönliches KI-Assistenzsystem",
    "title": "JARVIS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/runs": {
      "get": {
        "description": "Die eigenen Läufe, neueste zuerst.\n\nOhne Zugehörigkeitsprüfer, weil es hier nichts zu prüfen gibt: Die Abfrage\nkennt nur den Nutzer aus der Sitzung und hat keinen Parameter, über den ein\nanderer benannt werden könnte.",
        "operationId": "list_runs_runs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunView"
                  },
                  "title": "Response List Runs Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "description": "Legt einen Lauf an — für den angemeldeten Nutzer.\n\nDie Einstufung geschieht sofort und deterministisch. Sie bestimmt die\nDatenklasse des Laufs, und die ist die Obergrenze für alles, was später in\nihm geschieht: Sie nachträglich zu setzen hieße, sie nachträglich zu\nbehaupten.\n\nDer Lauf bleibt ``queued``. Ausgeführt wird er nicht — es gibt noch keine\nWerkzeuge und keinen Arbeiter, der ihn aufnähme.",
        "operationId": "create_run_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{run_id}": {
      "get": {
        "description": "Ein einzelner eigener Lauf — mit dem Plan und seinem jetzigen Stand.",
        "operationId": "read_run_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{run_id}/advance": {
      "post": {
        "description": "Führt den nächsten fälligen Schritt des Plans aus.\n\n**Der Plan bindet.** Bei ``POST /runs/{id}/steps`` nennt der Aufrufer das\nWerkzeug; hier nennt es der Plan. Das ist die engere Tür: Wer diesen\nEndpunkt benutzt, kann keinen Schritt überspringen und keinen einschieben,\nder nicht angekündigt war — und der Nutzer hat den Plan vorher gesehen.\n\n**Was diese Funktion tut, ist absichtlich wenig.** Sie löst den Lauf auf,\nprüft dessen Zugehörigkeit, übergibt an den ``RunAdvancer`` und übersetzt\ndessen Ausgang in eine Antwort. Der Ablauf selbst — Anspruch, Vorbereitung,\nWirkung, Festschreiben — steht in ``core/orchestrator/advance.py``.\n\nDas war nicht immer so, und der Umbau hat einen Anlass: An genau dieser\nGrenze sind zwei Sicherheitslücken kurz nacheinander entstanden, beide an\nder Reihenfolge *Anspruch → Wirkung → Festschreiben*. Sie war über eine\nRoutenfunktion verteilt und ließ sich nicht an einer Stelle überblicken.\nEin externer Prüfer hat beide gefunden und den Umbau vorgeschlagen.\n\n**Die Identität bleibt hier.** Wem ein Lauf gehört, entscheidet diese\nSchicht aus der Sitzung; der ``RunAdvancer`` bekommt den geladenen Lauf und\nhat keinen Parameter, mit dem sich ein fremder benennen ließe.",
        "operationId": "advance_run_runs__run_id__advance_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Advance Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{run_id}/steps": {
      "post": {
        "description": "Führt einen Werkzeugschritt aus — Glieder ⑤ bis ⑦ über HTTP.\n\n**Was dieser Endpunkt nicht entscheidet: irgendetwas.** Er löst den Lauf\nauf, prüft dessen Zugehörigkeit und übergibt an den Executor. Ob das\nWerkzeug laufen darf, entscheidet die Policy Engine; ob eine Bestätigung\nnötig ist, ebenfalls; ob der Grant gilt, das Gate; ob er noch nicht\nverbraucht ist, die Datenbank. Eine Prüfung hier wäre eine zweite Wahrheit.\n\n**Der Werkzeugname kommt vom Aufrufer, die Erlaubnis nicht.** Das ist der\nPfad für einen ausdrücklichen Befehl („lies mir die Datei X\") und\nzugleich der, den die Modellschleife später intern nimmt. Beide landen bei\nderselben Policy-Entscheidung — ein Vorschlag eines Modells trägt keine\nBerechtigung mit sich.\n\n**Fortschreiben mit Statusvergleich.** Gespeichert wird gegen den Status,\nder beim Laden galt. Läuft parallel ein zweiter Schritt, gewinnt genau\neiner; der andere bekommt 409 statt eines überschriebenen Laufs.",
        "operationId": "execute_step_runs__run_id__steps_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Step",
        "tags": [
          "runs"
        ]
      }
    }
  }
}